src/api/tests/requirements-test.txt
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
httpx==0.25.2
websockets==12.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
PyJWT==2.8.0
fastapi==0.110.0
uvicorn==0.27.1
pydantic-settings==2.2.1
